amazon/ec2: add user_data_gzip boolean
Allow users to specify if they'd like their user_data gzip'd.
Unfortunately, it's rather difficult to implement this as a filter due
to the resulting binary data and the handling of it at that level.

Futher, most people end up wanting to base64 the result of the gzip'd
user_data, but boto ends up doing that for you so there isn't room for a
gzip_and_base64 filter (which would work perfectly in ansible).

This option strikes a nice balance between the options.